feat: added the support for the dynamic update of the zone
