+++ config.yml
@@ -6,7 +6,7 @@ log:
   # The log level. This can be one of the following: trace, debug, info, warn, error, or off.
   level: debug
   # Enable the metrics.
-  enable_metrics: true
+  enable_metrics: false
   # Enable thread ID in logs.
   enable_thread_id: true
   # Log on stderr.
